Added all WIP Files
All files moved from internal/local repo to new public github based repo. Originally this script was a local project that I was working on in between help desk tickets where i work but due to time constraints the script project was cancelled and i was allowed to scrub all company based information and open source the work.

Within the reference folder there is a .vb file that is being used as the basis of this script. The original script was a VBScript that wasnt friendly for use by other helpdesk team members for on demand usage